Added wait, get results, use on next hit
So I have it now obtain results from the initial hit,verify hits
(approve/reject), generate the vote hit based on those results, obtain
the vote hit results and verify those. Then add the best choice to the
story. The posting to wordpress part that John added works with this as
well.